remove tests/reference from the regular build
Because it had a SConscript file, it was getting picked up
by BasicSConstruct.initialize and built as part of an
"install".  This is not desirable, as it requires gfortran.
Renaming the SConscript to SConstruct removes it from the
build, but you can still go into that directory and run
"scons" to build it.