Save raw variance and ignore weights instead of using a constant in Likelihood.
Previously, we used a determinant-based average for the weights when
not using per-pixel weights, which was never well-motivated, and got
in the way of using Likelihood objects to computing more principled
error estimates when fitting without using per-pixel weights.

This changeset will temporarily break the calculation of CModel
uncertainties (which apparently weren't well-tested enough to notice
the change).  That will be fixed in a future commit.