CModel: scale flux uncertainty to account for extrapolation
We attempt to extrapolate the flux to infinite, but we'd previously
estimated the uncertainty using only the pixels we use in the fit.
That can result in underestimated fluxes, so in this change we
scale that uncertainty by the ratio of the extrapolated flux to
the flux within the fit region only.