Use reasonable value for constant weights in Likelihood.
In likelihood-only fitting, it's acceptable to just set sigma=1
when you don't want different weights on each data point,
since sigma cancels out overall.  But it doesn't cancel out
overall when a prior is used, so we need to pick a value that's
at least approximately right; I've chosen the the arithmetic
mean of the variance.